# Optimize types.CertAuthority comparison