While fixing #19413, it was found that `CertAuthoritiesEquivalent` can take more than **5ms** per compare.

This change optimizes the comparison.

Before change:
```
$ go test -bench=BenchmarkCertAuthoritiesEquivalent -benchmem -run=^$
goos: darwin
goarch: arm64
pkg: github.com/gravitational/teleport/lib/services
BenchmarkCertAuthoritiesEquivalent/true-8         	     279	   4236727 ns/op	  323249 B/op	     375 allocs/op
BenchmarkCertAuthoritiesEquivalent/false-8        	     187	   6427850 ns/op	  494779 B/op	     505 allocs/op
```

With optimization:
```
$ go test -bench=BenchmarkCertAuthoritiesEquivalent -benchmem -run=^$
goos: darwin
goarch: arm64
pkg: github.com/gravitational/teleport/lib/services
BenchmarkCertAuthoritiesEquivalent/true-8         	   54924	     20904 ns/op	   18720 B/op	     187 allocs/op
BenchmarkCertAuthoritiesEquivalent/false-8        	   62005	     19032 ns/op	   18416 B/op	     157 allocs/op
```

It boosts `go test ./integration/ -run=TestIntegrations/Rotate`  on my mac for about 6 seconds (three tests in total). Pprof shows this compare function before this change was using 25%~30% of the CPU during the `RotateXXX` tests.

Note that, as a side effect, this does make the `RotateXXX` tests less flaky. To make the `RotateXXX` tests flakier (to reproduce the real problems), one can just add a 10ms delay in the compare.